requirements.txt
albumentations==1.1.0
matplotlib==3.2.1
numpy==1.13.3
opencv_python==4.2.0.34
pandas==1.1.5
Pillow==9.0.0
scikit_image==0.17.2
skimage==0.0
torch==1.10.1+cu113
torchvision==0.11.2+cu113
tqdm==4.46.0
